Make G_parser() only invoke UI if module has required arguments No need for "if (argc > 1 && G_parser(argc, argv))"
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@32756 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ display/d.measure/msurements.c
@@ -18,7 +18,7 @@ int measurements(int color1, int color2, int s_flag, int m_flag, int k_flag)
     int cur_screen_x, cur_screen_y;
     int screen_x, screen_y;
     struct Cell_head window;
-    int t, b, l, r;
+    double t, b, l, r;
 
     nalloc = 128;
     x = (double *)G_calloc(nalloc, sizeof(double));
